+++ src/Form/Related/Forms/HasMany.php
@@ -67,6 +67,13 @@ class HasMany extends Elements
 
                 $related->setAttribute($attribute, $value);
                 $element->setModel($related);
+
+                //for model hasmany->image, images
+                if($element instanceof Image){
+                    $request = new Request();
+                    $request->replace([$element->getPath() => $value]);
+                    $element->afterSave($request);
+                }
             }
         }
     }